+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to `@homebridge/ciao` will be documented in this file. This project tries to adhere to [Semantic Versioning](http://semver.org/).
 
-## v1.3.11 (Pending Release)
+## v1.3.11 (2026-08-13)
 
 ### Changes